build(cmake): add pt_warnings interface target

+++ include/pt/core/hittable_list.hpp
@@ -6,6 +6,7 @@
 #include "pt/math/random.hpp"
 #include "pt/math/ray.hpp"
 #include "pt/math/vec3.hpp"
+#include <cstddef>
 #include <memory>
 #include <vector>
 
@@ -38,7 +39,7 @@ public:
     [[nodiscard]] double pdf_value(const point3& origin, const vec3& direction) const override {
         if (objects.empty()) return 0.0;
 
-        const double weight = 1.0 / objects.size();
+        const double weight = 1.0 / static_cast<double>(objects.size());
         double sum{0.0};
 
         for (const auto& object : objects) {
@@ -51,7 +52,7 @@ public:
         if (objects.empty()) return vec3(0, 0, 0);
 
         const int count = static_cast<int>(objects.size());
-        return objects[random_int(0, count - 1)]->random(origin);
+        return objects[static_cast<std::size_t>(random_int(0, count - 1))]->random(origin);
     }
 
     void clear() {
